[fix] move add/remove logic to updateTeams function
